refactor (step5) : domain 패키지로 이동
