+++ src/main/kotlin/racingcar/domain/car/Winners.kt
@@ -1,4 +1,4 @@
-package racingcar.car
+package racingcar.domain.car
 
 @JvmInline
 value class Winners(val winners: List<Car>)
