+++ src/main/kotlin/racingcar/domain/engine/CarEngine.kt
@@ -1,4 +1,4 @@
-package racingcar.engine
+package racingcar.domain.engine
 
 interface CarEngine {
     fun execute(): Boolean
