+++ src/main/kotlin/racingcar/domain/random/RacingRandomGenerator.kt
@@ -1,4 +1,4 @@
-package racingcar.random
+package racingcar.domain.random
 
 class RacingRandomGenerator : RandomGenerator {
     override fun randomNum(max: Int): Int = (0..max).random()
