+++ src/main/kotlin/racingcar/domain/random/RandomGenerator.kt
@@ -1,4 +1,4 @@
-package racingcar.random
+package racingcar.domain.random
 
 interface RandomGenerator {
     fun randomNum(max: Int): Int
